refactor event_binner unit tests
